+++ src-tauri/src/dto/mod.rs
@@ -1,2 +1,3 @@
 pub mod fababox;
 pub mod device_management;
+pub mod system;